[api] Split Event::Project
Have every `Event::Project` subclass in a different file, as it is
normally done in Rails.
